评论区还没魔改

+++ article/views.py
@@ -3,26 +3,45 @@ from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from .forms import ArticlePostForm
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from comment.models import Comment
+import markdown
 
 
 def article_list(request):
     # 取出所有博客文章
     articles = ArticlePost.objects.all()
     # 需要传递给模板（templates）的对象
-    context = { 'articles': articles }
+    context = {'articles': articles}
     # render函数：载入模板，并返回context对象
     return render(request, 'article/list.html', context)
 
 
 def article_detail(request, id):
-    # 取出相应的文章
     article = ArticlePost.objects.get(id=id)
-    # 需要传递给模板的对象
-    context = { 'article': article }
+    md = markdown.Markdown(
+        extensions=[
+            # 包含 缩写、表格等常用扩展
+            'markdown.extensions.extra',
+            # 语法高亮扩展
+            'markdown.extensions.codehilite',
+            # 目录扩展
+            'markdown.extensions.toc',
+        ]
+    )
+
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
+    comments = Comment.objects.filter(article=id)
+
+    article.body = md.convert(article.body)
+
+    context = {'article': article, 'toc': md.toc, 'comments': comments}
     # 载入模板，并返回context对象
     return render(request, 'article/detail.html', context)
 
 
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     # 判断用户是否提交数据
     if request.method == "POST":
@@ -32,10 +51,7 @@ def article_create(request):
         if article_post_form.is_valid():
             # 保存数据，但暂时不提交到数据库中
             new_article = article_post_form.save(commit=False)
-            # 指定数据库中 id=1 的用户为作者
-            # 如果你进行过删除数据表的操作，可能会找不到id=1的用户
-            # 此时请重新创建用户，并传入此用户的id
-            new_article.author = User.objects.get(id=1)
+            new_article.author = User.objects.get(id=request.user.id)
             # 将新文章保存到数据库中
             new_article.save()
             # 完成后返回到文章列表
@@ -48,20 +64,24 @@ def article_create(request):
         # 创建表单类实例
         article_post_form = ArticlePostForm()
         # 赋值上下文
-        context = { 'article_post_form': article_post_form }
+        context = {'article_post_form': article_post_form}
         # 返回模板
         return render(request, 'article/create.html', context)
 
 
+@login_required(login_url='/userprofile/login/')
 def article_delete(request, id):
     # 根据 id 获取需要删除的文章
     article = ArticlePost.objects.get(id=id)
+    if request.user != article.author:
+        return HttpResponse("nmsl，你还想删？")
     # 调用.delete()方法删除文章
     article.delete()
     # 完成删除后返回文章列表
     return redirect("article:article_list")
 
 
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     """
     更新文章的视图函数
@@ -72,6 +92,9 @@ def article_update(request, id):
 
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+    if request.user != article.author:
+        return HttpResponse("改你妈哦")
+
     # 判断用户是否为 POST 提交表单数据
     if request.method == "POST":
         # 将提交的数据赋值到表单实例中
@@ -93,6 +116,6 @@ def article_update(request, id):
         # 创建表单类实例
         article_post_form = ArticlePostForm()
         # 赋值上下文，将 article 文章对象也传递进去，以便提取旧的内容
-        context = { 'article': article, 'article_post_form': article_post_form }
+        context = {'article': article, 'article_post_form': article_post_form}
         # 将响应返回到模板中
         return render(request, 'article/update.html', context)
